+++ lib/scss_lint/linter/duplicate_root.rb
@@ -1,23 +1,24 @@
 module SCSSLint
-  # Checks identical root selectors.
+  # Checks for identical root selectors.
   class Linter::DuplicateRoot < Linter
     include LinterRegistry
 
     def visit_root(node)
       # Root rules are evaluated per document, so use new hashes for eash file
-      @roots = Hash.new
+      @roots = {}
       yield # Continue linting children
     end
 
     def visit_rule(node)
-      # Check to see if we've seen this rule before
-      if @roots.has_key?(node.rule)
-        add_lint(node.line,
-                 "Root #{node.rule} already seen at #{@roots[node.rule]}")
+      if @roots[node.rule]
+        add_lint node.line,
+                 "Merge root rule `#{node.rule.join}` with identical " \
+                 "rule on line #{@roots[node.rule].line}"
       else
-        @roots[node.rule] = node.line
+        @roots[node.rule] = node
       end
-      return # Only check one level deep
+
+      # Don't yield so we only check one level deep
     end
   end
 end